+++ crates/koto/examples/hello_world.rs
@@ -1,0 +1,5 @@
+fn main() {
+    koto::Koto::default()
+        .compile_and_run("print 'Hello, World!'")
+        .unwrap();
+}